Moves creating baremetal until after all services are restarted and adds downloading vmlinuz and initrd.img.
There looks to be some errors in foreman-proxy log where Foreman is trying to
contact these services and failing when hosts are created.  Maybe moving
after these services are restarted will help.  The main reason for this
commit is to ensure that vmlinuz and initrd are downloaded correctly to
the /var/lib/tftpboot/boot/ dir.  Foreman sometimes doesn't download
them all the way and errors, but doesn't retry.