Changes the check to see if nodes are done with puppet provisioning
The old method seemed like a bad way to do things (check for a file on a
machine).  The new way is to query Foreman via it's REST API every
minute to see if each node is in "A" state.  A state means that puppet
has completed with no errors in its last run.  We wait up to 60 minutes.