Fixes issue where Ansible will not break loop on puppet check failure
Nodes have 60 minute timeouts to make sure puppet finishes on them and
are checked every minute to see if they are complete.  This is done in a
loop per node.  The problem is with Ansible, if a node fails to complete
during the 60 minute timeout it will still continue to do the 60 minute
timeout on the subsequent nodes.  This is giant waste of time since the
first node failing means the whole deloyment has failed.

This patch moves those loops into an Ansible python module where we can
control the logic.

JIRA: BGS-62